+++ project/apps/notify/src/app/email-subscriber/entity/email-subscriber.entity.ts
