+++ project/apps/notify/src/app/email-subscriber/repository/email-subscriber.repository.ts
@@ -3,8 +3,8 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 
 import { BaseMongoRepository } from '@project/shared/core';
-import { EmailSubscriberEntity } from './email-subscriber.entity';
-import { EmailSubscriberModel } from './email-subscriber.model';
+import { EmailSubscriberEntity } from '../entity/email-subscriber.entity';
+import { EmailSubscriberModel } from '../email-subscriber.model';
 
 @Injectable()
 export class EmailSubscriberRepository extends BaseMongoRepository<
